Package split: types-eslint
