+++ packages/parser-services/src/index.ts
@@ -1,6 +1,5 @@
 export * from './clear-caches';
 export { getCanonicalFileName } from './create-program/shared';
-export type * from './types';
 export { createProgramFromConfigFile as createProgram } from './create-program/useProvidedPrograms';
 export * from './withoutProjectParserOptions';
 export * from './parser';
