+++ packages/types-eslint/README.md
@@ -1,0 +1,3 @@
+# `@typescript-eslint/types-eslint`
+
+TODO
