+++ packages/types-eslint/project.json
@@ -1,0 +1,16 @@
+{
+  "name": "types-eslint",
+  "$schema": "../../node_modules/nx/schemas/project-schema.json",
+  "root": "packages/types-eslint",
+  "sourceRoot": "packages/types-eslint/src",
+  "projectType": "library",
+  "targets": {
+    "lint": {
+      "executor": "@nx/eslint:lint",
+      "outputs": ["{options.outputFile}"]
+    },
+    "test": {
+      "executor": "@nx/vite:test"
+    }
+  }
+}
