+++ packages/types-eslint/src/ESLint.ts
