+++ packages/types-eslint/src/Scope.ts
