+++ packages/types-eslint/src/index.ts
@@ -1,9 +1,11 @@
 export * from './AST';
+export * from './JSONSchema4';
 export * from './Config';
 export * from './ESLint';
 export * from './Linter';
 export * from './Parser';
 export * from './ParserOptions';
+export * from './ParserServices';
 export * from './Processor';
 export * from './Rule';
 export * from './RuleTester';
