+++ packages/types-eslint/vitest.config.mts
@@ -1,0 +1,26 @@
+import * as path from 'node:path';
+import { defineProject, mergeConfig } from 'vitest/config';
+
+import { vitestBaseConfig } from '../../vitest.config.base.mjs';
+import packageJson from './package.json' with { type: 'json' };
+
+const vitestConfig = mergeConfig(
+  vitestBaseConfig,
+
+  defineProject({
+    root: import.meta.dirname,
+
+    test: {
+      dir: path.join(import.meta.dirname, 'tests'),
+      name: packageJson.name.replace('@typescript-eslint/', ''),
+      root: import.meta.dirname,
+
+      setupFiles: [
+        './tests/test-utils/serializers/index.ts',
+        './tests/test-utils/custom-matchers/custom-matchers.ts',
+      ],
+    },
+  }),
+);
+
+export default vitestConfig;
