+++ packages/utils/package.json
@@ -66,6 +66,7 @@
     "@typescript-eslint/parser-services": "8.32.0",
     "@typescript-eslint/scope-manager": "8.32.0",
     "@typescript-eslint/types": "8.32.0",
+    "@typescript-eslint/types-eslint": "8.32.0",
     "@typescript-eslint/typescript-estree": "8.32.0"
   },
   "peerDependencies": {
