+++ packages/utils/src/ast-utils/eslint-utils/ReferenceTracker.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-namespace */
 import * as eslintUtils from '@eslint-community/eslint-utils';
 
-import type * as TSESLint from '../../ts-eslint';
+import type * as TSESLint from '@typescript-eslint/types-eslint';
 import type { TSESTree } from '../../ts-estree';
 
 const ReferenceTrackerREAD: unique symbol = eslintUtils.ReferenceTracker.READ;
