+++ packages/utils/src/ast-utils/eslint-utils/astUtilities.ts
@@ -1,6 +1,6 @@
 import * as eslintUtils from '@eslint-community/eslint-utils';
 
-import type * as TSESLint from '../../ts-eslint';
+import type * as TSESLint from '@typescript-eslint/types-eslint';
 import type { TSESTree } from '../../ts-estree';
 
 /**
