+++ packages/utils/src/ast-utils/eslint-utils/scopeAnalysis.ts
@@ -1,6 +1,7 @@
+import type * as TSESLint from '@typescript-eslint/types-eslint';
+
 import * as eslintUtils from '@eslint-community/eslint-utils';
 
-import type * as TSESLint from '../../ts-eslint';
 import type { TSESTree } from '../../ts-estree';
 
 /**
