+++ packages/utils/src/index.ts
@@ -1,7 +1,8 @@
+// TODO: Should this be deprecated? Probably not, since it's convenient, right?
+export * as TSESLint from '@typescript-eslint/types-eslint';
+
 export * as ASTUtils from './ast-utils';
 
 export * as ESLintUtils from './eslint-utils';
-export * as JSONSchema from './json-schema';
-export * as TSESLint from './ts-eslint';
 export * from './ts-estree';
 export * as TSUtils from './ts-utils';
